feat: moved consent proof handling to consent list load instead
Updated consent list to handle consent proofs
